Entrega a product owner

+++ client/src/pages/users/ListAllUsersPage.jsx
@@ -1,18 +1,32 @@
 import { useContext, useEffect, useState } from 'react';
 import { AuthContext } from '../../contexts/AuthContext.jsx';
-import { useDocumentTitle } from '../../hooks/index.js';
-import toast from 'react-hot-toast';
+import { useNavigate } from 'react-router-dom';
 
+import toast from 'react-hot-toast';
+import Swal from 'sweetalert2';
+
+import { useDocumentTitle } from '../../hooks/index.js';
 const { VITE_API_URL } = import.meta.env;
 
 //////
 
 const ListAllUsersPage = () => {
     useDocumentTitle('Gestión de usuarios'); // Título de pestaña
-    const { authToken } = useContext(AuthContext);
+    const { authToken, isAdmin } = useContext(AuthContext);
     const [users, setUsers] = useState([]);
+    const [filter, setFilter] = useState('all'); // Estado para el filtro de usuarios
+    const navigate = useNavigate();
 
     useEffect(() => {
+        // Si el usuario que trata de acceder a la página no es administrador, lo redirigimos automáticamente a la homepage
+        if (!isAdmin()) {
+            toast.error('No tienes permisos para realizar esa acción', {
+                id: 'listalluserspage',
+            });
+            navigate('/');
+            return;
+        }
+
         const fetchUsers = async () => {
             try {
                 const res = await fetch(
@@ -36,10 +50,11 @@ const ListAllUsersPage = () => {
                 // Hace que no se llame infinitamente
                 if (!compareUsers(body.data.users)) setUsers(body.data.users);
             } catch (err) {
-                toast.error(err.message);
+                toast.error(err.message, { id: 'listalluserspage' });
             }
         };
 
+        // Compara los usuarios actuales con los nuevos para evitar recargar innecesariamente
         const compareUsers = (newUsers) => {
             if (!users) {
                 return false;
@@ -48,15 +63,14 @@ const ListAllUsersPage = () => {
             if (users?.length !== newUsers?.length) return false;
 
             for (const user of users) {
-                //para cada usuario del state
-                //buscamos un ususraio con id equivalente en newUsers. Si no lo hay, devolvemos false
+                // Para cada usuario del state, buscamos uno con ID equivalente en newUsers.
                 const newUser = newUsers.find(
                     (newUser) => newUser.id === user.id,
                 );
 
                 if (!newUser) return false;
 
-                // si el usuraio de newUsers tiene una fecha de modificación distinta, devolvemos false
+                // Si el usuario tiene una fecha de modificación distinta, devolvemos false
                 if (user.updatedAt !== newUser.updatedAt) return false;
             }
 
@@ -64,20 +78,57 @@ const ListAllUsersPage = () => {
         };
 
         fetchUsers();
-    }, [authToken, users]);
+    }, [authToken, users, isAdmin, navigate]);
 
     const handleRemoveUser = async (user) => {
         try {
-            const userId = user.id;
+            // Comprobamos que no se está tratando de eliminar a un administrador
+            if (user.role === 'administrador') {
+                await Swal.fire({
+                    title: 'Acción no permitida',
+                    text: 'No se puede eliminar a un administrador.',
+                    icon: 'error',
+                });
+                return;
+            }
 
-            const res = await fetch(`${VITE_API_URL}/users/delete/${userId}`, {
-                method: 'DELETE',
-                headers: { Authorization: authToken },
+            // Confirmación con sweetalert2.
+            const resultDelete = await Swal.fire({
+                title: '¿Estás seguro de que quieres eliminar a este usuario?',
+                text: `Has seleccionado al usuario ${user.firstName} ${user.lastName}, con alias ${user.username}.`,
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#FF3333',
+                cancelButtonColor: '#22577A',
+                confirmButtonText: 'Sí, eliminar a este usuario',
+                cancelButtonText: 'Cancelar',
             });
-            const body = await res.json();
 
-            if (body.status === 'error') throw new Error(body.message);
-            toast.success(body.message);
+            // Si el administrador confirma, procedemos con la eliminación.
+            if (resultDelete.isConfirmed) {
+                const userId = user.id;
+
+                const res = await fetch(
+                    `${VITE_API_URL}/users/delete/${userId}`,
+                    {
+                        method: 'DELETE',
+                        headers: { Authorization: authToken },
+                    },
+                );
+                const body = await res.json();
+
+                if (body.status === 'error') throw new Error(body.message);
+
+                // Mostramos confirmación de eliminación con SweetAlert
+                await Swal.fire({
+                    title: 'Usuario eliminado',
+                    text: 'El usuario ha sido desactivado en la base de datos.',
+                    icon: 'success',
+                });
+
+                // Forzamos que la página se refresque automáticamente para actualizar la lista
+                window.location.reload();
+            }
         } catch (err) {
             toast.error(err.message, { id: 'alluserspage' });
         }
@@ -86,51 +137,88 @@ const ListAllUsersPage = () => {
     const handleButtonClick = async (user) => {
         try {
             if (user.active) {
-                if (
-                    !confirm(
-                        `¿Está seguro de que desea eliminar al usuario ${user.username}?`,
-                    )
-                )
-                    return;
                 handleRemoveUser(user);
             } else {
-                const res = await fetch(
-                    `${import.meta.env.VITE_API_URL}/users/addOrganizer/${user.id}`,
-                    {
-                        method: 'PUT',
-                        headers: {
-                            Authorization: authToken,
-                            'Content-Type': 'application/json',
-                        },
-                        body: JSON.stringify(user),
-                    },
-                );
-
-                const body = await res.json();
-
-                if (body.status === 'error') {
-                    throw new Error(body.message);
-                }
-
-                toast.success('Usuario activado correctamente', {
-                    id: 'alluserspage',
+                // Confirmación con sweetalert2.
+                const resultValidate = await Swal.fire({
+                    title: 'Usuario pendiente de validación',
+                    text: `Has seleccionado al usuario ${user.firstName} ${user.lastName}, con alias ${user.username}.`,
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#FF3333',
+                    cancelButtonColor: '#22577A',
+                    confirmButtonText: 'Validar a este usuario',
+                    cancelButtonText: 'Cancelar',
                 });
 
-                // Forzamos que la página se refresque automáticamente
-                window.location.reload();
+                // Si el administrador confirma, procedemos con la eliminación.
+                if (resultValidate.isConfirmed) {
+                    const res = await fetch(
+                        `${import.meta.env.VITE_API_URL}/users/addOrganizer/${user.id}`,
+                        {
+                            method: 'PUT',
+                            headers: {
+                                Authorization: authToken,
+                                'Content-Type': 'application/json',
+                            },
+                            body: JSON.stringify(user),
+                        },
+                    );
+
+                    const body = await res.json();
+
+                    if (body.status === 'error') {
+                        throw new Error(body.message);
+                    }
+
+                    toast.success('Usuario activado correctamente', {
+                        id: 'alluserspage',
+                    });
+
+                    // Forzamos que la página se refresque automáticamente para actualizar la lista
+                    window.location.reload();
+                }
             }
         } catch (err) {
             toast.error(err.message, { id: 'alluserspage' });
         }
     };
 
+    // Función que filtra usuarios
+    const filteredUsers = users.filter((user) => {
+        if (filter === 'active') {
+            return user.active;
+        } else if (filter === 'pending') {
+            return !user.active;
+        } else {
+            return true;
+        }
+    });
+
     return (
         <div className="w-full px-4 lg:px-24 py-8">
             <h1 className="text-header-big text-center ml-0">
                 Gestión de Usuarios
             </h1>
+
+            {/* Filtro de usuarios */}
+            <div className="mt-4 flex justify-end">
+                <label className="mr-4 font-semibold my-auto align-middle">
+                    Filtrar por estado:
+                </label>
+                <select
+                    className="p-2 border rounded"
+                    value={filter}
+                    onChange={(e) => setFilter(e.target.value)}
+                >
+                    <option value="all">Todos</option>
+                    <option value="active">Activado</option>
+                    <option value="pending">Pendiente</option>
+                </select>
+            </div>
+
             <div className="overflow-x-auto mt-8">
-                <table className="min-w-full bg-white shadow-md rounded-lg overflow-hidden">
+                <table className="min-w-full shadow-md rounded-lg overflow-hidden">
                     <thead>
                         <tr className="bg-azuloscuro text-blanco">
                             <th className="text-left py-3 px-4 uppercase font-semibold text-sm">
@@ -143,6 +231,9 @@ const ListAllUsersPage = () => {
                                 Email
                             </th>
                             <th className="text-left py-3 px-4 uppercase font-semibold text-sm">
+                                Rol
+                            </th>
+                            <th className="text-left py-3 px-4 uppercase font-semibold text-sm">
                                 Estado
                             </th>
                             <th className="text-left py-3 px-4 uppercase font-semibold text-sm">
@@ -150,15 +241,16 @@ const ListAllUsersPage = () => {
                             </th>
                         </tr>
                     </thead>
-                    <tbody className="text-gray-700">
-                        {users.length > 0 ? (
-                            users.map((user) => (
+                    <tbody>
+                        {filteredUsers.length > 0 ? (
+                            filteredUsers.map((user) => (
                                 <tr key={user.id} className="border-b">
                                     <td className="py-3 px-4">{user.id}</td>
                                     <td className="py-3 px-4">
                                         {user.firstName + ' ' + user.lastName}
                                     </td>
                                     <td className="py-3 px-4">{user.email}</td>
+                                    <td className="py-3 px-4">{user.role}</td>
                                     <td className="py-3 px-4">
                                         {user.active ? 'Activado' : 'Pendiente'}
                                     </td>
@@ -187,7 +279,7 @@ const ListAllUsersPage = () => {
                             ))
                         ) : (
                             <tr>
-                                <td colSpan="5" className="text-center py-4">
+                                <td colSpan="6" className="text-center py-4">
                                     No hay usuarios disponibles.
                                 </td>
                             </tr>
